viscosity added

+++ project_module_a/inc/Variables.hpp
@@ -1,7 +1,8 @@
 #ifndef DEF_VARIABLE_HPP
 #define DEF_VARIABLE_HPP
 
-using Real = double;
+using Real = float;
+
 using Dim = int;
 
 #endif // DEF_VARIABLE_HPP
